+++ include/exiv2/xmp_exiv2.hpp
@@ -138,8 +138,8 @@ namespace Exiv2 {
         // Todo: Remove this method from the baseclass
         //! The Exif typeSize doesn't make sense here. Return 0.
         size_t typeSize() const override;
-        long count() const override;
-        long size() const override;
+        size_t count() const override;
+        size_t size() const override;
         std::string toString() const override;
         std::string toString(long n) const override;
         long toLong(long n =0) const override;
