+++ src/xmp.cpp
@@ -222,12 +222,12 @@ namespace Exiv2 {
         return 0;
     }
 
-    long Xmpdatum::count() const
+    size_t Xmpdatum::count() const
     {
         return p_->value_.get() == 0 ? 0 : p_->value_->count();
     }
 
-    long Xmpdatum::size() const
+    size_t Xmpdatum::size() const
     {
         return p_->value_.get() == 0 ? 0 : p_->value_->size();
     }
